Refactor sunrise and sunset data handling by moving logic to a new module and updating local storage checks
