+++ index.html
@@ -7,6 +7,7 @@
     <link rel="stylesheet" href="CurrentAndHourly.css">
 </head>
 <script type="module" src="./js/vertexAI.js"></script>
+<script type="module" src="./js/sun_data.js"></script>
 <body>
     <div class="container">
         <div class="header">
