+++ js/current_hourly.js
@@ -1,4 +1,3 @@
-import { getCurrentDate, checkLocalStorage, checkAge } from "./app.js";
 document.addEventListener("DOMContentLoaded", function () {
     let currentUnit = 'F'; // Default unit is Fahrenheit
 
@@ -61,33 +60,7 @@ document.addEventListener("DOMContentLoaded", function () {
         updateTemperatureDisplay();
     }
 
-    // Function to fetch sunrise and sunset times
-    async function fetchSunriseSunset() {
-        //const { latitude, longitude } = currentWeather;
-        let latitude = localStorage.getItem("currentLocation").split(",")[0];
-        let longitude = localStorage.getItem("currentLocation").split(",")[1];
-        const url = `https://api.sunrise-sunset.org/json?lat=${latitude}&lng=${longitude}&formatted=0`;
 
-        try {
-            const response = await fetch(url);
-            const data = await response.json();
-
-            // Convert sunrise and sunset from UTC to local time
-            currentWeather.sunrise = convertToLocalTime(data.results.sunrise);
-            currentWeather.sunset = convertToLocalTime(data.results.sunset);
-            localStorage.setItem("sunData", getCurrentDate() + "|" + currentWeather.sunrise + "|" + currentWeather.sunset);
-            // Update sunrise and sunset times
-            document.getElementById("sunrise-time").innerText = `Sunrise: ${currentWeather.sunrise}`;
-            document.getElementById("sunset-time").innerText = `Sunset: ${currentWeather.sunset}`;
-        } catch (error) {
-            console.error("Error fetching sunrise and sunset times:", error);
-        }
-    }
-
-    function convertToLocalTime(isoTime) {
-        const date = new Date(isoTime);
-        return date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
-    }
 
     document.getElementById("current-temperature").innerText = `${currentWeather.temperature}°F`;
     document.getElementById("current-description").innerText = currentWeather.description;
@@ -98,12 +71,5 @@ document.addEventListener("DOMContentLoaded", function () {
     // Add event listener to the toggle button
     document.getElementById("toggleButton").addEventListener('click', toggleTemperature);
 
-    // Fetch sunrise and sunset times when the page loads and update the temperature display
-    async function localStorageDataChecks() {
-        if (!checkLocalStorage("daily","sunData") || checkAge("sunData")) {
-            fetchSunriseSunset();
-        }
-    }
-    localStorageDataChecks();
     updateTemperatureDisplay();
 });
